[android] Refactoring. Move Review class from SponsoredHotel class.
